move pages/Admin/templates/cw3/ReviewCW3Configurer to pages/Admin/Review/Templates
